Upgrade to JasperFx/Marten/Weasel rc; Saga.Version is permanently int
Bumps the critter-stack dependencies to their rc lines (holding Polecat,
which has no rc yet):
  - JasperFx family       -> 2.0.0-rc.2 / 5.0.0-rc.2
  - Marten / AspNetCore / Newtonsoft -> 9.0.0-rc.1
  - Weasel family         -> 9.0.0-rc.1
  - Polecat               -> held at 4.0.0-alpha.10

JasperFx 2.0 rc split numeric versioning into two interfaces:
  - IRevisioned.Version   : int   (optimistic-concurrency revision)
  - ILongVersioned.Version: long  (event-stream version)

Accordingly:
  - Saga.Version reverts from long back to int (permanent design decision)
    so sagas implement IRevisioned directly without a shadow override.
  - MartenIntegration resolves the saga revision member via
    GetProperty(nameof(Saga.Version), typeof(int)) to match.
  - LoadDocumentFrame keeps the generated expectedSagaRevision as a long
    because IDocumentOperations.UpdateRevision(entity, long) takes a long;
    the int Saga.Version + 1 widens into it. Comment corrected.
  - Event-sourced aggregate test types that track a long stream version
    (SideEffects1, GpStreamAggregate, GSimpleAggregate) now implement
    ILongVersioned instead of IRevisioned.

Test/sample compile fixes for rc namespace relocations:
  - IRevisioned: Marten.Metadata -> JasperFx
  - TenancyStyle: Marten.Storage -> JasperFx.MultiTenancy
  - Identity/IdentityAttribute: Marten.Schema -> JasperFx
  - ISoftDeleted: Marten.Metadata -> JasperFx.Metadata
  - PolecatTests: alias [Identity] to Wolverine.Polecat.IdentityAttribute
    to disambiguate from the lifted JasperFx.IdentityAttribute

Full wolverine.slnx builds clean (0 warnings, 0 errors) on net9.0/net10.0.

+++ src/Wolverine/Saga.cs
@@ -30,12 +30,13 @@ public abstract class Saga
     /// For saga providers that support this, this is a version of the saga to help enforce optimistic concurrency
     /// protections. This value is the current version that is stored by saga storage and will
     /// be incremented upon save.
-    /// Widened to <see cref="long"/> in 6.0 to align with Marten's
-    /// <c>Marten.Metadata.IRevisioned.Version</c> (which is <see cref="long"/>
-    /// from Marten 9.0.0-alpha.2 onward), so sagas can implement <c>IRevisioned</c>
-    /// without a shadow override.
+    /// Typed as <see cref="int"/> to align with <c>JasperFx.IRevisioned.Version</c>
+    /// (an <see cref="int"/>), so sagas can implement <c>IRevisioned</c> directly
+    /// without a shadow override. (JasperFx 2.0 rc split versioning into
+    /// <c>IRevisioned</c> = <see cref="int"/> and <c>ILongVersioned</c> = <see cref="long"/>;
+    /// sagas use the <see cref="int"/> revision.)
     /// </summary>
-    public long Version { get; set; }
+    public int Version { get; set; }
 }
 
 #endregion
